AMP updates
Deal with youtube videos more appropriately
Improve Structure Data
Extensionless posts should look at configuration (hard coded)
Removed pop3 info
Improved amp CSS
