+++ .gitignore
@@ -199,3 +199,4 @@ FakesAssemblies/
 
 # Visual Studio 6 workspace options file
 *.opt
+source/newtelligence.DasBlog.Web/content/**
