+++ source/newtelligence.DasBlog.Web.Core/newtelligence.DasBlog.Web.Core.csproj
@@ -162,7 +162,7 @@
     <Compile Include="Amp\AmpifyPageFilter.cs" />
     <Compile Include="Amp\BaseAmpProcess.cs" />
     <Compile Include="Amp\IProcessTag.cs" />
-    <Compile Include="Amp\ProcessIFrame.cs" />
+    <Compile Include="Amp\ProcessYouTube.cs" />
     <Compile Include="Amp\ProcessImage.cs" />
     <Compile Include="Amp\ProcessTwitter.cs" />
     <Compile Include="ArchiveMonthsList.cs">
