+++ source/newtelligence.DasBlog.Web/SiteConfig/site.config
@@ -46,7 +46,7 @@
   <AlwaysIncludeContentInRSS>true</AlwaysIncludeContentInRSS>
   <EntryTitleAsLink>true</EntryTitleAsLink>
   <ObfuscateEmail>true</ObfuscateEmail>
-  <NotificationEMailAddress>scott@hanselman.com</NotificationEMailAddress>
+  <NotificationEMailAddress>test@test.com</NotificationEMailAddress>
   <SendCommentsByEmail>true</SendCommentsByEmail>
   <SendReferralsByEmail>false</SendReferralsByEmail>
   <SendTrackbacksByEmail>true</SendTrackbacksByEmail>
@@ -68,7 +68,7 @@
   <UseUserCulture>true</UseUserCulture>
   <ShowItemDescriptionInAggregatedViews>true</ShowItemDescriptionInAggregatedViews>
   <EnableClickThrough>false</EnableClickThrough>
-  <ExtensionlessUrls>true</ExtensionlessUrls>
+  <ExtensionlessUrls>false</ExtensionlessUrls>
   <EnableAggregatorBugging>false</EnableAggregatorBugging>
   <DisplayTimeZoneIndex>4</DisplayTimeZoneIndex>
   <AdjustDisplayTimeZone>false</AdjustDisplayTimeZone>
@@ -77,11 +77,6 @@
   <BinariesDir>~/content/binary/</BinariesDir>
   <SmtpServer>localhost</SmtpServer>
   <EnablePop3>false</EnablePop3>
-  <Pop3Server>mail2.orcsweb.com</Pop3Server>
-  <Pop3Username>weblog@hanselman.com</Pop3Username>
-  <Pop3Password>cheeseman</Pop3Password>
-  <Pop3SubjectPrefix>Tweak:</Pop3SubjectPrefix>
-  <Pop3Interval>600</Pop3Interval>
   <Pop3InlineAttachedPictures>true</Pop3InlineAttachedPictures>
   <Pop3InlinedAttachedPicturesThumbHeight>150</Pop3InlinedAttachedPicturesThumbHeight>
   <ApplyContentFiltersToWeb>true</ApplyContentFiltersToWeb>
@@ -94,7 +89,7 @@
   <XSSUpstreamInterval>7200</XSSUpstreamInterval>
   <ContentFilters />
   <CrosspostSites>
-    <CrosspostSite profileName="LongHornBlogs" hostName="www.longhornblogs.com" port="80" endpoint="shanselman/services/metablogapi.aspx" username="shanselman" password="tweak" apitype="metaweblog" />
+    <CrosspostSite profileName="LongHornBlogs" hostName="www.longhornblogs.com" port="80" endpoint="" username="" password="" apitype="metaweblog" />
   </CrosspostSites>
   <Pop3DeleteAllMessages>true</Pop3DeleteAllMessages>
   <Pop3LogIgnoredEmails>true</Pop3LogIgnoredEmails>
