+++ source/newtelligence.DasBlog.Web/themes/amp/homeTemplate.blogtemplate
@@ -9,9 +9,197 @@
 		<script async custom-element="amp-form" src="https://cdn.ampproject.org/v0/amp-form-0.1.js"></script>
 		<script async custom-element="amp-iframe" src="https://cdn.ampproject.org/v0/amp-iframe-0.1.js"></script>
 		<script async custom-element="amp-twitter" src="https://cdn.ampproject.org/v0/amp-twitter-0.1.js"></script>
+		<script async custom-element="amp-youtube" src="https://cdn.ampproject.org/v0/amp-youtube-0.1.js"></script>
+
+<style amp-custom="">
+
+/* Fonts */
+
+@font-face {
+    font-family: 'Merriweather';
+    src: url('https://s1.wp.com/i/fonts/merriweather/merriweather-regular-webfont.woff2') format('woff2'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-regular-webfont.woff') format('woff'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-regular-webfont.ttf') format('truetype'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-regular-webfont.svg#merriweatherregular') format('svg');
+    font-weight: 400;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'Merriweather';
+    src: url('https://s1.wp.com/i/fonts/merriweather/merriweather-italic-webfont.woff2') format('woff2'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-italic-webfont.woff') format('woff'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-italic-webfont.ttf') format('truetype'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-italic-webfont.svg#merriweatheritalic') format('svg');
+    font-weight: 400;
+    font-style: italic;
+}
+
+@font-face {
+    font-family: 'Merriweather';
+    src: url('https://s1.wp.com/i/fonts/merriweather/merriweather-bold-webfont.woff2') format('woff2'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-bold-webfont.woff') format('woff'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-bold-webfont.ttf') format('truetype'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-bold-webfont.svg#merriweatherbold') format('svg');
+    font-weight: 700;
+    font-style: normal;
+}
+
+@font-face {
+    font-family: 'Merriweather';
+    src: url('https://s1.wp.com/i/fonts/merriweather/merriweather-bolditalic-webfont.woff2') format('woff2'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-bolditalic-webfont.woff') format('woff'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-bolditalic-webfont.ttf') format('truetype'), url('https://s1.wp.com/i/fonts/merriweather/merriweather-bolditalic-webfont.svg#merriweatherbold_italic') format('svg');
+    font-weight: 700;
+    font-style: italic;
+}
+
+.amp-db-content, .amp-db-title-bar div {
+    max-width: 600px;
+    margin: 0 auto;
+}
+
+body {
+    font-family: 'Merriweather', Serif;
+    font-size: 16px;
+    line-height: 1.8;
+    background: #fff;
+    color: #3d596d;
+    padding-bottom: 100px;
+}
+
+.amp-db-content {
+    padding: 4px;
+    overflow-wrap: break-word;
+    word-wrap: break-word;
+    font-weight: 400;
+    color: #3d596d;
+}
+
+.amp-db-title {
+    margin: 20px 0 0 0;
+    font-size: 30px;
+    line-height: 1.258;
+    font-weight: 700;
+    color: #2e4453;
+}
+
+/* UI Fonts */
+.amp-db-meta,
+nav.amp-db-title-bar,
+.db-caption-text {
+    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen-Sans", "Ubuntu", "Cantarell", "Helvetica Neue", sans-serif;
+    font-size: 15px;
+}
+
+.amp-db-meta {
+    margin-bottom: 16px;
+}
+
+/* Post meta data */
+ul.amp-db-meta {
+    padding: 24px 0 0 0;
+    margin: 0 0 24px 0;
+}
+
+    ul.amp-db-meta li {
+        list-style: none;
+        display: inline-block;
+        margin: 0;
+        line-height: 24px;
+        white-space: nowrap;
+        overflow: hidden;
+        text-overflow: ellipsis;
+        max-width: 300px;
+    }
+
+        ul.amp-db-meta li:before {
+            content: "\2022";
+            margin: 0 8px;
+        }
+
+        ul.amp-db-meta li:first-child:before {
+            display: none;
+        }
+
+.amp-db-meta,
+.amp-db-meta a {
+    color: #4f748e;
+}
+
+
+
+/* Post Header */
+nav.amp-db-title-bar {
+    background: #1f3b70;
+    padding: 0 16px;
+}
+
+nav.amp-db-title-bar div {
+    line-height: 54px;
+    color: #fff;
+}
+
+nav.amp-db-title-bar a {
+    color: #fff;
+    text-decoration: none;
+}
+
+/* Post Content */
+blockquote {
+    padding: 16px;
+    margin: 8px 0 24px 0;
+    border-left: 2px solid #87a6bc;
+    color: #4f748e;
+    background: #e9eff3;
+}
+
+blockquote p:last-child {
+    margin-bottom: 0;
+}
+
+
+amp-carousel {
+    background: #000;
+}
+
+amp-iframe,
+amp-youtube,
+amp-instagram,
+amp-vine {
+    background: #f3f6f8;
+}
+
+amp-carousel > amp-img > img {
+    object-fit: contain;
+}
+
+.amp-db-enforced-sizes {
+    max-width: 100%;
+}
+
+.amp-db-unknown-size img {
+    object-fit: contain;
+}
+
+p,
+ol,
+ul,
+figure {
+    margin: 0 0 24px 0;
+}
+
+a,
+a:visited {
+    color: #8e1303;
+}
+
+a:hover,
+a:active,
+a:focus {
+    color: #8e1303;
+}
+
+</style>
+
 	</head>
-	<body>
+	<body class="amp-mode-touch amp-mode-mouse amp-mode-keyboard-active">
+		<nav class="amp-db-title-bar">
+        		<div>
+				<a href="http://www.poppastring.com">
+					poppastring - Mark Downie
+				</a>
+			</div>
+		</nav>
 		<%bodytext%>
 	</body>
-	<footer>Accelerate Mobile Project (AMP) theme created by <a href="http://www.poppastring.com/">Mark Downie</a></footer>
 </html>
