+++ source/newtelligence.DasBlog.Web/themes/amp/itemTemplate.blogtemplate
@@ -1,8 +1,19 @@
-<div class="itemBoxStyle">
-	<div class="itemTitleStyle">
-		<%ItemTitle%>
-	</div>
-	<div class="itemBodyStyle">
-		<%ItemBody%>
-	</div>
+<div class="amp-db-content">
+    <h1 class="amp-db-title"><%ItemTitle%></h1>
+
+    <ul class="amp-db-meta">
+        <li>
+            <%FormattedWhenBare("MMMM d, yyyy H:mm")%>
+        </li>
+        <li>
+            <span class="screen-reader-text">Categories:</span>
+            <%categoryLinks%>
+        </li>
+        <li>
+            <span class="screen-reader-text">Share on:</span>
+            <%TwitterShareLink("Twitter", "")|custommacro%>, <%FaceBookShareLink("Facebook", "")|custommacro%> and <%GooglePlusShareLink("Google+", "")|custommacro%>
+        </li>
+    </ul>
+        
+    <%ItemBody%>
 </div>
